Manage the scenario test types
The scenario syntax could define also the type of the test that will
be generated. It exists four types :
* ui
* unit
* plugin
* manual

The syntax is :
Scenario: Name [Parent Story] @ui
A UI test will be generated for the scenario.
If no test type notation is added :
Scenario: Name [Parent Story]
The test type is considered as to be set to manual.